npu-monitor-tool: refactor code in preparation for Prometheus
Move the main telemetry implementation of NPU monitoring into
core/main.py. This is in preparation for prometheus exporting.

Also, take advantage to remove PTL arch check w.r.t. to npu
memory utilization reading as it should depend on VPU driver
published sysfs npu_memory_utilization directly. In another words,
it is the role of VPU driver to publish the sysfs if it supports it.

The tool responsibility is on formatting the value for presentation
based on the raw value or print 'NA' if sysfs is not valid.

This refactoring preserves the original functionality.
